xx-3.2.2 some fix
