+++ Weather_app_mvvm_test/View/AddCityVC.swift
@@ -62,8 +62,9 @@ class AddCityVC: UIViewController, UITableViewDelegate, UITableViewDataSource, U
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        
+     
         let searchResult = searchResults[indexPath.row]
+        
         let cell = UITableViewCell(style: .subtitle, reuseIdentifier: nil)
         cell.contentView.backgroundColor = .secondarySystemBackground
         cell.backgroundColor = .secondarySystemBackground
