+++ Weather_app_mvvm_test/View/Extensions/UIImageView+Extention.swift
@@ -1,0 +1,15 @@
+import UIKit
+
+extension UIImageView {
+    func loadImage(withUrl url: URL) {
+        DispatchQueue.global().async { [weak self] in
+            if let imageData = try? Data(contentsOf: url) {
+                if let image = UIImage(data: imageData) {
+                    DispatchQueue.main.async {
+                        self?.image = image
+                    }
+                }
+            }
+        }
+    }
+}
